Added an action to find object changes
This find changes object is more accurate than the changed date or the
object state.

this action will actually compare values.  By default changing an
attribute, and then changing the value back will mark the object as
'changed'.  This Java also does a value compare and will only mark it as
changed if there actually is a different value.